+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 - **incompatible change** updated resource name for `google_dns_policy` on the `net-vpc` module
+- added support for VPC-SC Ingress Egress policies on the `vpc-sc` module
 
 ## [4.8.0] - 2021-05-12